+++ R/utils_dt.R
@@ -169,3 +169,17 @@ recycle <- function(x, size) {
 recycle_eval <- function(expr, data = NULL, env= rlang::caller_env(), size){
   recycle(rlang::eval_tidy(expr, data = data, env = env),size = size)
 }
+#' @noRd
+changed_objects <- function(obj){
+  ## name <- rlang::eval_tidy(rlang::as_name(rlang::enquo(obj)))
+  oo <- ls(envir=.GlobalEnv)
+  mem <- data.table::data.table(mem = lapply(oo, function(x)  do.call(data.table::address,list(rlang::sym(x))) ) %>% unlist(), names = oo)
+
+  loc <- data.table::address(force(obj))
+  changed <- mem[mem ==loc,]$names
+  if(length(changed)>1){
+    message("The following objects have been changed in place: ", paste0(changed,sep =", "))
+  } else {
+    message(changed, " has been changed in place.")
+  }
+}