+++ NEWS.md
@@ -3,7 +3,8 @@
     - removed dependency with `stringr`
     - changed internal structure of `eeg_lst`, the segments table is a data.table rather than a tibble
   - **Potential breaking change**: when segments with repeated columns are joined, they use data.table notation (`i...`) rather than dplyr notation. This might affect the `join` family and `eeg_segment`.
-  
+ -  filters have more options (including IIR and by reference!).
+ 
 # eeguana 0.1.5.9000
   - Bugs: 
     - weird dplyr filter behavior fixed  [#136](https://github.com/bnicenboim/eeguana/issues/136)